Clarify how to join a Github Classroom team and get a repo URL
In her review of Pull Request #17, @kklamberty suggested some
clarifications to the bit about joining a Github classroom team
and how to get the URL for a repository. Hopefully these help with this.

I'm not thrilled about this text, though; I feel like it's beginning to
blur together a not-very-good tutorial on Github Classroom with checking
out a project. Do we need to back up and revisit the structure, perhaps
talking about how to use Classroom in the write-up? Or are we putting
too much effort into writing up something that we'll just walk them
through in class?