Support more complex names
The initial test only support single-word, ASCI names with no
punctuation (e.g., hyphens). This uses [a regex from StackOverflow](https://stackoverflow.com/questions/15805555/java-regex-to-validate-full-name-allow-only-spaces-and-letters)
to support a broader range of names.

It also turns out that I totally forgot to include an assertion when
checking that the first line was correct. This also fixes that.