Update README.md to fix issue #25 and issue #26, change Github to GitHub
This commit updates the README file to clarify a couple of elements of the lab that have caused confusion in the past including making it clear that you need to log in to GitHub using a username and password rather than the default option of using a token. The directions also erroneously said to update main when it should say to update `generateOutput`, so that is fixed here as well. Finally, this commit updates the appearance of Github to be GitHub so that our text is consistent and matches what one would find on GitHub.