Update HellosTest.java to fix issue #20, bring back test from last year
testLineStructure() was added last year and was still visible in the master branch, but was not in the develop branch. This commit brings that test here and makes it work with the version of testing we are using. It further updates the test so that the greeting does not need to be 'Hello!', but does still need to have the '!'