Switch from Travis to GitHub Action badges
This switches from a Travis-CI badge to a GitHub badge. I also added a note about fixing the badges to point to the particular semester repo instead of the "master" repo, complete with a "broken" badge from <Shields.io> at the top to catch everyone's attention.

This closes #36. I'm also going to say this closes #12 as well since the warning should hopefully be enough of a reminder.