Proof and update down to "Merging your branch into `main`"
This provides a pretty thorough (and time consuming) proof-reading and updating down to the start of the "Merging your branch into `main`" section.

I'm seriously considering modifying the end game to use pull requests instead of hand merging. GitHub's support for PRs has gotten fancier and niftier, including warnings when your branch is behind `main` and giving you a push-button way to merge `main` into your PR branch.

We definitely want them to be using pull requests and code reviews through the semester, so maybe we should just start with that.

That will probably slow things down, but I'm thinking that this is going to become a two lab-session event in a purely on-line COVID world this semester anyway.