Add `settings.yml`
I've installed the Settings App for this organization, and this is a my stab at a settings file that hopefully covers the key settings that we want to set.

I suspect we won't what this does or doesn't do until we actually merge it into `master`, and we may not _really_ know what happens until we create the assignment in GitHub Classroom to see what happens there.