Update info on setting `git` email and name
It seems that GitKraken may do all this for the students. I've added a note to that effect, and had them check that `user.email` and `user.name` are configured correctly. If they are, they can skip the rest of that section; if not they should press on and set them.

I did suggest that they might want to do this on their lab accounts, though, since they'll need that eventually (assuming we return to the lab in the fall).