Add notes about turning on/off requirements for tests and reviews
We probably want branch protection on for the second half, but we _have_ to have of off for the first half. Thus at a minimum we need a reminder note for the faculty member that they need to turn things off and on again.

I'm thinking we might be able to semi-automate this with the settings files (`setting.yml`).