Remove all the notes in README about running tests in VS Code
This removes all mention of the "built-in" test runner in VS Code and instead goes over opening an VS Code terminal window and running tests from there.

Since our build system is so Gradle dependent, it's not clear that in later labs students will be able to actually run the JUnit tests from the integrated test runner in VS Code, so it seemed to make sense to just drop all that from here.