Add note to README about the need to authenticate when you `push`
There was nothing about the need to authenticate when you run `git push`. This turns out to be kind of a complicated mess at the moment as GitHub is changing their policies, etc., so I just of dodged the problem for now and essentially just said "do the obvious things and ask questions if you're stuck".