Document using PRs in the README
This is a really big change, which switches from doing all the merging in GitKraken to doing the merging in GitHub. I'm not entirely sure how I feel about this, and we're running out of time to make a decision.

I've left a big chunk of old text at the end that talks about merge conflicts and doing merges "by hand" in GitKraken. I don't think that's necessary here, but it's potentially useful information and maybe needs to be saved in a separate document?