Fix some mistakes in branch protection doc
I had misunderstood what it would mean to have an "empty" branch restriction (where nothing was checked), and @floogulinc clarified that this would still prevent anyone from pushing directly to `main`, which would break the first half of the lab entirely!

This fixes that, and adds a restriction on who can dismiss pull request reviews.