Switch to Google style guide
The previous `checkstyle.xml` was based on the Sun style guide. There's also [a `checkstyle.xml` that is Google styleguide](https://checkstyle.org/google_style.html), and arguably these days Google's opinion counts for more than Sun's (or even Oracle's), so thought I'd switch to the Google version.

In the next commit I'll see what needs to be fixed so that we "pass".