Add Markdown link check to GitHub Actions
Adds a `link-check.yaml` GitHub Actions workflow file that uses the `gaurav-nelson/github-action-markdown-link-check` action to check our links.

I didn't turn on the `modified-files-only` feature, but that may be something we want to do down the road.

I also didn't add scheduled runs, but it might be worth having this run every month or so rather than just have it happen on push events.