Add link check config to ignore GitHub Action badge links
The GitHub Action badge links fail to check here. I'm guessing that's the use relative paths is what breaks things, but I'm not sure.

I hope this sets up a config file that says to ignore URLs that start with "../../actions/workflows" to avoid us having to worry about those.