Move documentation into `docs/` directory
This moves the `FACULTY_BRANCH_PROTECTION_SETTINGS.md` and `MERGE_CONFLICTS.md` files into the `docs/` directory as suggested in #76.

I also fixed all the links to point to the new location, and added a few more links so that students/readers would be more likely to find those resources.