A bazillion updates to the docs for S22
Most of this was driven by the need to switch (back) to the class being in the lab instead of remote like in S21. While I went through it, though, I made a bunch of other (mostly minor) changes along the way, fixing typos and trying to clarify things.

I also added a bunch of images to try to illustrate various points.