Mention possible warnings about Java installation paths
Adding the `foojay` plugin that downloads the correct version if it's not on the system generates a bunch of warnings when you run Gradle as it looks for the "right" version of Java. This adds a note telling the students that they can safely ignore these.

I also used the `details` HTML tag to provide more information in a "hidden" block in case they want to learn more about what's going on here.