Run CodeQL on all pushes and PRs
For some reason this was set up to only (I think) run if there was a push onto `main`, which basically never happens directly because `main` is protected.

I also changed the cron from running once a week to running once a month. Given that we're not really worried about security issues on this repo, and it sits dormant for months at a stretch, it seemed to make more sense to have it run less often.