Attempt to suppress unused type warning
SonarCloud (and JUnit5 best practices) suggest that test classes should be package visible. Unfortunately that causes CodeQL to complain that we have an unused type, which is a potential security problem.

I'm trying to resolve that by suppressing the `unused-reference-type` warning, but I don't know if that will actually work. We'll see.