Replace a "raw" `for` loop with a `for-each` loop
This addresses a warning from DeepSource about the fact that we were using a "raw" `for` loop when we should be using a `for-each` style loop.

I had hoped that I could also replace the other `for` loop, but that can't be easily done since we're looping over two parallel arrays and need the index for the error message.