Add `merge_group` to all the workflow files
There's been some kind of change in the world, and GitHub now says that we're supposed to have `merge_group` as a trigger for all our workflows. This adds that.

https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue#triggering-merge-group-checks-with-github-actions

I don't think we've been having problems by _not_ having this, but if it is indeed necessary (or going to be necessary), I figured I'd add it.

I'm a little worried about their comments about needing to configure 3rd party checks. It's possible that things like DeepSource will break (or fail to ever return a check) on these things?

It's possible that we're "safe" because we're triggering everything on `push`, which basically causes pretty much everything to be checked. I'm not at all sure, though.